python packaging, make tool installable as command line tool via `pip install git+https://github.com/MEGA-GO/Mega-Go.git`
