+++ .gitignore
@@ -1,12 +1,3 @@
 
-tair.gaf
-go-basic.obo
-goa_uniprot_all.gaf
-goa_uniprot_all.gaf.gz
 .ipynb_checkpoints/similarty_metrics-checkpoint.ipynb
 .Rproj.user
-compare_goa.csv
-GO terms.csv
-uniprot-reviewed_yes.tab
-uniprot-sp.tab
-.ipynb_checkpoints/similarty_metrics-checkpoint.ipynb
