+++ requirements-dev.txt
@@ -1,0 +1,4 @@
+# Dependencies, with versions explicitly declared for development and testing
+goatools==0.9
+# Packages that are not dependencies but used for building or testing
+pylint
