+++ testdata/example_input-compare_goa.csv
@@ -1,0 +1,23 @@
+set1,set2,visualize,class
+GO:0030170,GO:0070279,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0030170,GO:0070279/chart?showKey=true",one level higher
+GO:0019842,GO:0036094,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0019842,GO:0036094/chart?showKey=true",one level higher
+GO:0030170,GO:0043168,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0030170,GO:0043168/chart?showKey=true",one level higher
+GO:0030170,GO:0043167,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0030170,GO:0043167/chart?showKey=true",two levels higher
+GO:0030170,GO:0036094,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0030170,GO:0036094/chart?showKey=true",two levels higher
+GO:0070279,GO:0036094,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0070279,GO:0036094/chart?showKey=true",two levels higher
+GO:0030170,GO:0003674,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0030170,GO:0003674/chart?showKey=true",four levels higher
+GO:0050662,GO:0001844,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0050662,GO:0001844/chart?showKey=true",two unrelated bindings (MF vs BP)
+GO:1903876,GO:0097524,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:1903876,GO:0097524/chart?showKey=true",two unrelated bindings (MF vs CC)
+GO:1903876,GO:0048290,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:1903876,GO:0048290/chart?showKey=true",two unrelated bindings (MF vs BP)
+GO:0043168;GO:0050662,GO:0050662;GO:0070279,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0043168,GO:0050662,GO:0050662,GO:0070279/chart?showKey=true",same level bindings (unrel)
+GO:0048038;GO:0050662,GO:0043295;GO:0031419,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0048038,GO:0050662,GO:0043295,GO:0031419/chart?showKey=true",same level bindings (unrel)
+GO:0001913;GO:0042267,GO:0070942;GO:0090634,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0001913,GO:0042267,GO:0070942,GO:0090634/chart?showKey=true",same level bindings (unrel)
+GO:0005488,GO:0048037;GO:0036094,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0005488,GO:0048037,GO:0036094/chart?showKey=true",two children of parent
+GO:0098581,GO:0032490;GO:0098543,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0098581,GO:0032490,GO:0098543/chart?showKey=true",two children of parent
+GO:0050789,GO:0030155;GO:0031341,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0050789,GO:0030155, GO0031341/chart?showKey=true",two children of parent (BP)
+GO:0048037;GO:0036094,GO:1901363;GO:0097159,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0048037,GO:0036094,GO:1901363,GO:0097159/chart?showKey=true","two children of binding, vs two other children of binding"
+GO:0001906;GO:0009758,GO:0019740;GO:0098743,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0001906,GO:0009758,GO:0019740,GO:0098743/chart?showKey=true","two children of biological process, vs two other children of biological process"
+GO:0051712,,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0051712,/chart?showKey=true","one report, other not (BP)"
+GO:0097159,GO:0097159;GO:0005488;GO:0003674,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0097159,GO:0097159,GO:0005488,GO:0003674/chart?showKey=true",report specific; vs specific + parents
+GO:0030170,,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0030170,/chart?showKey=true","one report, other not"
+GO:0030170,GO:0003924,"https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/GO:0030170,GO:0003924/chart?showKey=true",unrelated (GTPase activity) vs binding (PPB) activity
